Adds a minimum viable level of authentication
Allows config to specify a level of permission required to be able to
interact with webhooks.  You can set this to sysadmin to only allow
sysadmins to create hooks, you can set this to admin to only someone
with organization management permissions, or you can set to editor which
will allow anyone who has admin or editor for an organization.  By
default this still allows for anyone to create a webhook (if they're
logged in).